requirements.txt
streamlit
langchain
pinecone-client
sentence-transformers
pypdf 
langchain-pinecone
langchain-community 
langchain-huggingface 
huggingface_hub 
unstructured
requests